refactor(prom links): reorganize code
Prepare for consistent behaviour of prominent links rendering regardless of context, in particular fallback behaviour that requires access to the DB (not available in "site" TS project, see https://github.com/owid/owid-grapher/pull/1073)
