+++ site/blocks/ProminentLink.tsx
@@ -146,40 +146,6 @@ export class ProminentLink extends React.Component<{
     }
 }
 
-export const renderAuthoredProminentLinks = ($: CheerioStatic) => {
-    $("block[type='prominent-link']").each((_, el: CheerioElement) => {
-        const $block = $(el)
-        const href = $block.find("link-url").text()
-        const url = Url.fromURL(href)
-
-        const style = $block.attr("style")
-        const title = $block.find("title").text()
-        const content = $block.find("content").html()
-        const image =
-            $block.find("figure").html() ||
-            (url.isGrapher
-                ? `<img src="${BAKED_GRAPHER_EXPORTS_BASE_URL}/${url.pathname
-                      ?.split("/")
-                      .pop()}.svg" />`
-                : null)
-
-        const rendered = ReactDOMServer.renderToStaticMarkup(
-            <div className="block-wrapper">
-                <ProminentLink
-                    href={href}
-                    style={style}
-                    title={title}
-                    content={content}
-                    image={image}
-                />
-            </div>
-        )
-
-        $block.after(rendered)
-        $block.remove()
-    })
-}
-
 export const hydrateProminentLink = (
     globalEntitySelection?: SelectionArray
 ) => {
